Centralize SignalBlocker helper
